refactor(p2): extract ChatListSupport into sibling file (stage 1 of 4)
Moves the tail helpers segment (KDoc L1738 + decls L1741-2009) from
ChatList.kt into the new sibling ChatListSupport.kt. Includes 15 funs:
extractMatchingSnippet, 4 LazyListState extensions (isNearListEnd,
isAtTimelineBottom, markdownPrewarmTexts, buildLazyItemMessageIndexMap),
List<Int?>.firstLazyIndexForMessage, ChatMessageVirtualItem.isAdjacentMarkdownChild,
TimelineSelectableMessageItem composable, MessageNode.markdownPrewarmTexts,
Conversation.actionSuggestionTexts, 5 String-extension helpers in the
action-suggestion chain, and buildHighlightedText.

Byte-equal: bodies and signatures unchanged.

Visibility widenings (6 sibling p→i): extractMatchingSnippet,
LazyListState.markdownPrewarmTexts, ChatMessageVirtualItem.isAdjacentMarkdownChild,
TimelineSelectableMessageItem, Conversation.actionSuggestionTexts, buildHighlightedText.
(5 sibling-internal helpers — MessageNode.markdownPrewarmTexts and 4
String-action helpers — kept private; already-internal isNearListEnd /
isAtTimelineBottom / buildLazyItemMessageIndexMap / firstLazyIndexForMessage
unchanged.)

Visibility widenings (5 coord p→i, for sibling access to shared consts):
MarkdownPrewarmBeforeItems, MarkdownPrewarmAfterItems, MarkdownPrewarmMaxTexts,
ActionOptionLineRegex, ActionOptionCuePhrases.

Coordinator strip: 2293 -> 2010 (-283 with import cleanup).
Imports trimmed: 10 orphans (Checkbox, AnnotatedString, SpanStyle,
buildAnnotatedString, withStyle, UIMessagePart, Assistant,
AssistantAffectScope, replaceRegexes, ChatMessageVirtualItem) — all
verified body refs only in extracted range.

Pre-existing dead imports preserved verbatim.

Baseline 3/467 unit tests preserved.
